requirements.txt
python-dotenv==0.16.0
requests==2.25.1
suntime=1.2.5
urllib3==1.26.4
